Adjust the blog structure
I had trouble making a blog page where the landing page for the directory is accessible from the navbar, but also doesn't appear as an item in a foler in the drop down.

My workaround is to have the "blog" page be a standalone page that links to articles in a separate "posts" directory. Perhaps not the cleanest way to do it, but it will do for now.